*Added rude reply to messages containing 'groovy' *SparksieBOT will now remove messages beginning with a groovy command if not in the channel "groovybitch-corner" and then ping the invoker in that channel *SparksieBOT will now remove any messages from Groovy that appear outside of the channgel "groovybitch-corner" *Restricted access to !say. Currently spark.c#7001 is hardcoded as the only accepted user. Others will get a mouthy response.

+++ cogs/Functions.py
@@ -87,8 +87,11 @@ class Functions(commands.Cog):
     
     @commands.command()
     async def say(self, ctx, *args:str):
-        await ctx.message.delete()
-        await ctx.channel.send(' '.join(args))
+        if str(ctx.message.author) == 'spark.c#7001':
+            await ctx.message.delete()
+            await ctx.channel.send(' '.join(args))
+        else:
+            await ctx.channel.send('Nice try. I\'m not your mouthpiece anymore!')
 
     @commands.command()
     async def sleepy(self, ctx):
